requirements.txt
# docs.google.com automation
pygame
pyautogui
pyperclip
soundfile
ffmpeg-python